+++ plugins/modules/ds8000_volume_mapping.py
@@ -2,9 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # Copyright (C) 2021 IBM CORPORATION
-# Author(s): Matan Carmeli <matan.carmeli7@gmail.com>
-#
-# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+# Apache License, Version 2.0 (see https://opensource.org/licenses/Apache-2.0)
 
 from __future__ import absolute_import, division, print_function
 
@@ -12,16 +10,18 @@ __metaclass__ = type
 
 DOCUMENTATION = r'''
 ---
-author: Matan Carmeli (@matancarmeli7)
 module: ds8000_volume_mapping
-short_description: Manage DS8000 volume mapping to hosts.
+short_description: Manage DS8000 volume mapping to hosts
 description:
   - Manage DS8000 volume mapping to hosts.
+version_added: "1.0.0"
+author: Matan Carmeli (@matancarmeli7)
 options:
   name:
     description:
       - The name of the DS8000 host to work with.
     type: str
+    required: true
   state:
     description:
     - Specify the state the DS8000 volume mapping should be in.
@@ -38,36 +38,39 @@ options:
     description:
       - The volume name that you want to map to a host.
       - Notice that different volumes sometimes have the same volume name, so it will map all of them.
-      - To use a specific volume, use C(volume_id)
+      - To use a specific volume, use I(volume_id)
     type: str
+notes:
+  - Supports C(check_mode).
 extends_documentation_fragment:
   - ibm.ds8000.ds8000.documentation
 '''
 
 EXAMPLES = r'''
-- name: Ensure that a host exists in the storage
+- name: Ensure that a volume is mapped to a host in the storage
   ibm.ds8000.ds8000_volume_mapping:
     hostname: "{{ ds8000_host }}"
     username: "{{ ds8000_username }}"
     password: "{{ ds8000_password }}"
     name: host_name_test
     state: present
+    volume_id: "0000"
 
-- name: Ensure that a host does not exist in the storage
+- name: Ensure that a volume is not mapped to a host in the storage
   ibm.ds8000.ds8000_volume_mapping:
     hostname: "{{ ds8000_host }}"
     username: "{{ ds8000_username }}"
     password: "{{ ds8000_password }}"
     name: host_name_test
     state: absent
+    volume_id: "0000"
 '''
 
 RETURN = r''' # '''
 
-import json
-from ansible_collections.ibm.ds8000.plugins.module_utils.ds8000 import (Ds8000ManagerBase, ds8000_argument_spec)
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.common.text.converters import to_native
+from ansible_collections.ibm.ds8000.plugins.module_utils.ds8000 import Ds8000ManagerBase, ds8000_argument_spec
 
 
 class VolumeMapper(Ds8000ManagerBase):
@@ -96,17 +99,15 @@ class VolumeMapper(Ds8000ManagerBase):
                 return
         name = self.params['name']
         try:
-            self.client.map_volume_to_host(host_name=name,
-                                           volume_id=volume_id)
+            if not self.module.check_mode:
+                self.client.map_volume_to_host(host_name=name, volume_id=volume_id)
             self.changed = True
         except Exception as generic_exc:
             self.failed = True
             self.module.fail_json(
-                msg="Failed to map volume id {volume_id} to {host_name} host on DS8000 storage. "
-                "ERR: {error}".format(
-                    volume_id=volume_id,
-                    host_name=name,
-                    error=to_native(generic_exc)))
+                msg="Failed to map volume id {volume_id} to host {host_name} on the DS8000 storage system. "
+                "ERR: {error}".format(volume_id=volume_id, host_name=name, error=to_native(generic_exc))
+            )
 
     def _get_volumes_mapping_on_specific_host(self):
         name = self.params['name']
@@ -124,17 +125,15 @@ class VolumeMapper(Ds8000ManagerBase):
         name = self.params['name']
         lun_id = volume_map.lunid
         try:
-            self.client.unmap_volume_from_host(host_name=name,
-                                               lunid=lun_id)
+            if not self.module.check_mode:
+                self.client.unmap_volume_from_host(host_name=name, lunid=lun_id)
             self.changed = True
         except Exception as generic_exc:
             self.failed = True
             self.module.fail_json(
-                msg="Failed to unmap volume id {volume_id} from {host_name} host on DS8000 storage."
-                " ERR: {error}".format(
-                    volume_id=volume_map.volume,
-                    host_name=name,
-                    error=to_native(generic_exc)))
+                msg="Failed to unmap volume id {volume_id} from host {host_name} on the DS8000 storage system."
+                " ERR: {error}".format(volume_id=volume_map.volume, host_name=name, error=to_native(generic_exc))
+            )
 
 
 def ensure_volume_mapping_state(volume_id, module, volume_mapper):
@@ -148,31 +147,29 @@ def ensure_volume_mapping_state(volume_id, module, volume_mapper):
 def main():
     argument_spec = ds8000_argument_spec()
     argument_spec.update(
-        name=dict(type='str'),
-        state=dict(type='str',
-                   default='present', choices=['absent', 'present']),
+        name=dict(type='str', required=True),
+        state=dict(type='str', default='present', choices=['absent', 'present']),
         volume_id=dict(type='str'),
-        volume_name=dict(type='str'))
+        volume_name=dict(type='str'),
+    )
 
     module = AnsibleModule(
         argument_spec=argument_spec,
         required_one_of=[
             ['volume_name', 'volume_id'],
         ],
-        supports_check_mode=False,
+        supports_check_mode=True,
     )
 
     volume_mapper = VolumeMapper(module)
 
-    if volume_mapper.verify_ds8000_object_exist('name', volume_mapper.client.get_hosts()):
+    if volume_mapper.verify_ds8000_object_exist(volume_mapper.client.get_host, host_name=module.params['name']):
         if module.params.get('volume_name'):
-            volume_ids = volume_mapper.get_volume_ids_from_name(
-                module.params['volume_name'])
+            volume_ids = volume_mapper.get_volume_ids_from_name(module.params['volume_name'])
             for volume_id in volume_ids:
                 result = ensure_volume_mapping_state(volume_id, module, volume_mapper)
         else:
-            result = ensure_volume_mapping_state(
-                module.params['volume_id'], module, volume_mapper)
+            result = ensure_volume_mapping_state(module.params['volume_id'], module, volume_mapper)
     else:
         volume_mapper.failed = True
         result = {'changed': volume_mapper.changed, 'failed': volume_mapper.failed}