+++ CHANGELOG.rst
@@ -1,5 +1,18 @@
-Will be updated by antsibull-changelog. Do not edit this manually!
+========================
+ibm.ds8000 Release Notes
+========================
 
-See https://github.com/ansible-community/antsibull-changelog/blob/main/docs/changelogs.rst for information on how to use antsibull-changelog.
+.. contents:: Topics
 
-Check out ``changelogs/config.yaml`` for its configuration. You need to change at least the ``title`` field in there.
+
+v1.0.0
+======
+
+New Modules
+-----------
+
+- ds8000_host - Manage DS8000 hosts
+- ds8000_host_port - Manage host ports for a DS8000 host
+- ds8000_volume - Manage DS8000 volumes
+- ds8000_volume_info - Return basic info on DS8000 volumes
+- ds8000_volume_mapping - Manage DS8000 volume mapping to hosts